requirements.txt
# Datagroom MCP Server (Python) - production dependencies
# Python >= 3.10

fastmcp>=2.0.0,<3
pydantic>=2.0.0,<3
httpx>=0.25.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pymongo>=4.0.0
